Split out statsd from lib/analytics
